Readjusted the effect of magic cards.
- change: 緑の上位カードの効果を変更した。
- change: 黒の魔法カードのターゲットを拡大した。
- change: 魔法服による効果上昇の対象を拡大。
- fix: 赤のバッジの効果使用時のクラッシュを修正。

+++ main/java/defeatedcrow/hac/magic/client/MagicCircleRenderer.java
@@ -1,5 +1,7 @@
 package defeatedcrow.hac.magic.client;
 
+import defeatedcrow.hac.core.client.base.DCFoodModelBase;
+import defeatedcrow.hac.magic.proj.EntityMobBarrier;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.client.renderer.OpenGlHelper;
 import net.minecraft.client.renderer.entity.Render;
@@ -7,8 +9,6 @@ import net.minecraft.client.renderer.entity.RenderManager;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
-import defeatedcrow.hac.core.client.base.DCFoodModelBase;
-import defeatedcrow.hac.magic.proj.EntityMobBarrier;
 
 @SideOnly(Side.CLIENT)
 public class MagicCircleRenderer extends Render<EntityMobBarrier> {
@@ -34,14 +34,13 @@ public class MagicCircleRenderer extends Render<EntityMobBarrier> {
 		GlStateManager.translate((float) x, (float) y + height, (float) z);
 		GlStateManager.enableBlend();
 		GlStateManager.disableLighting();
-		GlStateManager.tryBlendFuncSeparate(GlStateManager.SourceFactor.SRC_ALPHA,
-				GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA, GlStateManager.SourceFactor.ONE,
-				GlStateManager.DestFactor.ZERO);
+		GlStateManager
+				.tryBlendFuncSeparate(GlStateManager.SourceFactor.SRC_ALPHA, GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA, GlStateManager.SourceFactor.ONE, GlStateManager.DestFactor.ZERO);
 
 		int i = 15728880;
 		int j = i % 65536;
 		int k = i / 65536;
-		OpenGlHelper.setLightmapTextureCoords(OpenGlHelper.lightmapTexUnit, (float) j, (float) k);
+		OpenGlHelper.setLightmapTextureCoords(OpenGlHelper.lightmapTexUnit, j, k);
 
 		GlStateManager.color(2.0F, 2.0F, 2.0F, 0.75F);
 		this.bindTexture(tex);
